+++ week07/jungsukwoo-crud-api/src/main/java/practice/jungsukwoocrudapi/member/entity/Member.java
@@ -5,7 +5,7 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import practice.jungsukwoocrudapi.post.Post;
+import practice.jungsukwoocrudapi.post.entity.Post;
 
 import java.util.ArrayList;
 import java.util.List;
