+++ week07/jungsukwoo-crud-api/src/main/java/practice/jungsukwoocrudapi/post/controller/PostController.java
@@ -1,8 +1,7 @@
-package practice.jungsukwoocrudapi.post;
+package practice.jungsukwoocrudapi.post.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
+import practice.jungsukwoocrudapi.post.service.PostService;
 
 @RequiredArgsConstructor
 public class PostController {
