+++ week07/jungsukwoo-crud-api/src/main/java/practice/jungsukwoocrudapi/post/controller/dto/request/CreatePostRequest.java
@@ -1,6 +1,7 @@
-package practice.jungsukwoocrudapi.post;
+package practice.jungsukwoocrudapi.post.controller.dto.request;
 
 import lombok.Getter;
+import practice.jungsukwoocrudapi.post.entity.Post;
 
 @Getter
 public class CreatePostRequest {
