+++ week07/jungsukwoo-crud-api/src/main/java/practice/jungsukwoocrudapi/post/service/PostService.java
@@ -1,7 +1,8 @@
-package practice.jungsukwoocrudapi.post;
+package practice.jungsukwoocrudapi.post.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import practice.jungsukwoocrudapi.post.repository.PostRepository;
 
 @Service
 @RequiredArgsConstructor
